removed unused jobs code [2395]
